+++ src/pywmapi/auth/api.py
@@ -62,14 +62,16 @@ def signin(
             "email": email,
             "password": password,
             "device_id": device_id,
-            "auth_type": auth_type,
+            "auth_type": auth_type.value,
         },
         headers={"X-CSRFToken": csrf_token},
         cookies={"JWT": jwt},
     )
     check_wm_response(res)
     user = User.from_dict(res.json()["payload"]["user"])
-    return Session(res.cookies["JWT"], csrf_token, user, ws_platform, on_message=ws_on_message)
+    return Session(
+        res.cookies["JWT"], csrf_token, user, ws_platform, on_message=ws_on_message
+    )
 
 
 def restore(email: str) -> None: