+++ src/pywmapi/exceptions/defs.py
@@ -8,7 +8,9 @@ __all__ = [
 class WMError(Exception):
     """Base exception for pywmapi"""
 
-    def __init__(self, status_code: int, error_msg: Optional[str], raw_error: Exception) -> None:
+    def __init__(
+        self, status_code: int, error_msg: Optional[str], raw_error: Exception
+    ) -> None:
         super().__init__(status_code, error_msg, raw_error)
         self.status_code = status_code
         self.error_msg = error_msg